+++ light-spike/src/components/clock.rs
@@ -1,0 +1,12 @@
+use crate::prelude::*;
+
+pub trait Clock {
+    fn now(&self) -> Time;
+}
+
+pub struct SystemClock;
+impl Clock for SystemClock {
+    fn now(&self) -> Time {
+        Time::now()
+    }
+}
