+++ light-spike/src/prelude.rs
@@ -2,17 +2,18 @@ pub use std::time::Duration;
 
 pub use tendermint::time::Time;
 
-pub use crate::components::demuxer::*;
-pub use crate::components::fork_detector::*;
-pub use crate::components::io::*;
-pub use crate::components::scheduler::*;
-pub use crate::components::state::*;
-pub use crate::components::verifier::*;
-pub use crate::errors::*;
-pub use crate::operations::*;
-pub use crate::predicates::errors::*;
-pub use crate::predicates::VerificationPredicates;
-pub use crate::store::*;
-pub use crate::types::*;
-pub use crate::Never;
-pub use crate::{bail, ensure, postcondition, precondition};
+pub use crate::{
+    bail,
+    components::{
+        clock::*, demuxer::*, fork_detector::*, io::*, scheduler::*, state::*, verifier::*,
+    },
+    ensure,
+    errors::*,
+    operations::*,
+    postcondition, precondition,
+    predicates::errors::*,
+    predicates::VerificationPredicates,
+    store::*,
+    types::*,
+    Never,
+};
